backend-py/app/pyproject.toml

[tool.mypy]
python_version = "3.12"
exclude = ['.venv/']
strict = true
ignore_missing_imports = true
install_types = true
non_interactive = true

[tool.ruff]
# Same as Black.
line-length = 88

# Assume Python 3.11.
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN1",   # missing type self and cls
    "E999",   # syntax error
    "UP027",  # unpacked list comprehension
    "PT004",  # pytest missing fixture name underscore
    "PT005",  # pytest incorrect fixture name underscore
    "D1",     # undocumented
    "TD001",  # invalid todo tag
    "TD002",  # missing todo author
    "TD003",  # missing todo link
    "PD011",  # pandas use of dot values
    "B008",
]
unfixable = [
    "F401",   # unused import
    "F841",   # unused variable
    "ERA001", # commented out code
]

logger-objects = ["src.library.logger.LOGGER"]

[tool.ruff.isort]
combine-as-imports = true
known-first-party = ["musubi_restapi"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
split-on-trailing-comma = true

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
docstring-code-format = true

[lint.mccabe]
max-complexity = 3

[project]
name = "backend-py"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "orjson ==3.10.14",
    "fastapi",
    "starlette",
    "uvicorn[standard]",
    "python-multipart",
    "supabase",
    "strawberry-graphql",
    "pandas",
    "numpy",
    "scipy",
    "httpx",

    # Prisma
    "prisma",

    # Autogen
    "pyautogen",

    # OpenAI
    "openai",

    # LangChain
    "langchain",
    "langchain-community",
    "langchain-anthropic",
    "langchain-openai",
    "langchain-postgres",
    "langgraph",
    "langchainhub",

    # 画像
    "Pillow",

    # 画像生成AI Client
    "replicate",
    "fal-client",

    # runpod
    "runpod",

    # Modal
    "modal",

    # AI
    "torch",
    # WebRTC
    "aiortc",
    "livekit",
    "livekit-api",

    # Message Queue
    "kombu",
    "tembo-pgmq-python",

    # AWS
    # "boto3",

    # Google Cloud
    # "google-cloud-bigquery-data-exchange",

    # Azure
    # "azure-identity",

    # Snowflake
    # "snowflake-connector-python",

    # 音声
    "cartesia",

    # Logging
    "structlog",

    "pyyaml>=6.0.2",
    "oyaml>=1.0",
]

[dependency-groups]
dev = [
    # 画像生成AI
    "diffusers",
    "transformers",
    "sentencepiece",
    "protobuf",
    "accelerate",
    "compel",
    "torchao",
    "huggingface-hub",
]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"
